requirements.txt
attrs==17.4.0
click==6.7
daiquiri==1.3.0
